export idxs

+++ src/main/java/net/suteren/stardict/wiktionary2stardict/jpa/entity/LanguageCombinationEntity.java
@@ -1,4 +1,4 @@
 package net.suteren.stardict.wiktionary2stardict.jpa.entity;
 
-public record LanguageCombinationEntity(String from, String to, int count) {
+public record LanguageCombinationEntity(String from, String to, long count) {
 }
